+++ repology_client/constants.py
@@ -10,7 +10,7 @@ wrong?
 PACKAGE = "repology-client"
 
 #: Library version.
-VERSION = "0.2.0"
+VERSION = "0.3.0"
 
 #: Library homepage.
 HOMEPAGE = "https://repology-client.sysrq.in"